[#63] Renamed and simplified reducer names
