+++ src/reducers/playersListReducer.js
